refactor: rename settings_schemas to backoffice_schemas
