+++ src/core/infra/database/drizzle/schema.ts
@@ -3,4 +3,4 @@ export * from "./schemas/auth_schemas";
 export * from "./schemas/property_schemas";
 export * from "./schemas/stay_schemas";
 export * from "./schemas/finance_schemas";
-export * from "./schemas/settings_schemas";
+export * from "./schemas/backoffice_schemas";
